Fix a "bug" in Static Templates.
Groups in a static template file that were *not* marked as late activated would spawn into the world upon the _DATABASE:Spawn() call. So we manually force all groups in `RegisterSTMTable()` to be late activated. Users wishing for their static templates to spawn on mission start should call one of the "Spawn" methods made available in the FMS API.